backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
anthropic==0.31.0
pypdf2==3.0.1
python-docx==1.1.2
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.4.0
tenacity==8.5.0
structlog==24.4.0
